requirement.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.17.0
billiard==4.2.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
cssselect2==0.8.0
distro==1.9.0
Django==5.2
django-anymail==13.0
django-crontab==0.7.1
django-extensions==4.1
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.1
et_xmlfile==2.0.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
kombu==5.5.4
lxml==6.0.0
MarkupSafe==3.0.2
openai==1.78.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.9.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
pywin32==311
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
xhtml2pdf==0.2.17
